split request module
