+++ actix-http/src/lib.rs
@@ -40,7 +40,7 @@ mod helpers;
 mod http_message;
 mod message;
 mod payload;
-mod request;
+mod requests;
 mod response;
 mod response_builder;
 mod service;
@@ -58,9 +58,9 @@ pub use self::extensions::Extensions;
 pub use self::header::ContentEncoding;
 pub use self::http_message::HttpMessage;
 pub use self::message::ConnectionType;
-pub use self::message::{Message, RequestHead, RequestHeadType, ResponseHead};
+pub use self::message::{Message, ResponseHead};
 pub use self::payload::{Payload, PayloadStream};
-pub use self::request::Request;
+pub use self::requests::{Request, RequestHead, RequestHeadType};
 pub use self::response::Response;
 pub use self::response_builder::ResponseBuilder;
 pub use self::service::HttpService;
