+++ actix-http/src/requests/mod.rs
@@ -1,0 +1,7 @@
+//! HTTP requests.
+
+mod head;
+mod request;
+
+pub use self::head::{RequestHead, RequestHeadType};
+pub use self::request::Request;
